requirements.txt
Django==1.11.12
django-classy-tags==0.8.0
django-cms==3.4.6
django-formtools==2.1
django-sekizai==0.10.0
django-treebeard==4.2.0
djangocms-admin-style==1.2.7
mysqlclient==1.3.12
pytz==2017.3
djangocms_text_ckeditor
aldryn_style
aldryn-apphooks-config
aldryn_categories
aldryn_common
aldryn_newsblog
django-filer
djangocms_picture
djangorestframework
markdown
django-localflavor
django-smart-selects
django-widget-tweaks
phonenumbers
django-phonenumber-field
#django-recaptcha2
django-simple-captcha
django-crispy-forms
